# 00-PROJECTS/HAIDA/haida-main/dev/docs/API/openapi.yaml
openapi: 3.0.3
info:
  title: HAIDA API
  description: |
    HAIDA - Quality Assurance Platform with AI Integration

    A comprehensive testing and auditing platform that combines traditional QA automation
    with AI-powered insights for enterprise software quality assurance.

    ## Key Features
    - Multi-tenant architecture with RBAC
    - Automated testing execution and monitoring
    - AI-powered chat and assistance (Copilot Studio)
    - Comprehensive documentation management
    - Real-time analytics and reporting
    - Enterprise SSO integration (Microsoft Entra ID)
  version: 2.0.0
  contact:
    name: HAIDA Development Team
    email: hola@stayarta.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://haida.vercel.app/api/v2
    description: Production server
  - url: http://localhost:8000/api/v2
    description: Development server
security:
  - bearerAuth: []
  - entraAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated requests
    entraAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/{tenant}/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/{tenant}/oauth2/v2.0/token
          scopes:
            User.Read: Read user profile
            Files.Read.All: Read files
            Calendars.Read: Read calendar
            Mail.Read: Read mail
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2025-12-16T10:30:00Z"
    Email:
      type: string
      format: email
      example: "user@company.com"
...
    User:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        email:
          $ref: '#/components/schemas/Email'
        full_name:
          type: string
          example: "John Doe"
        avatar_url:
          type: string
          format: uri
        role:
          type: string
          enum: [owner, admin, editor, viewer]
        is_active:
          type: boolean
          default: true
        tenant_id:
          $ref: '#/components/schemas/UUID'
        preferences:
          type: object
          properties:
            theme:
              type: string
              enum: [light, dark]
              default: light
            notifications:
              type: object
              properties:
                email:
                  type: boolean
                  default: true
                push:
                  type: boolean
                  default: true
            dashboard_layout:
              type: string
              default: default
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
...
paths:
...
  /auth/login:
    post:
      summary: Local user login
      description: Authenticate user with email/password
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  $ref: '#/components/schemas/Email'
                password:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token
                  refresh_token:
                    type: string
                    description: JWT refresh token
                  token_type:
                    type: string
                    example: bearer
                  expires_in:
                    type: integer
                    description: Token expiration in seconds
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/register:
    post:
      summary: User registration
      description: Register a new user account
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, full_name]
              properties:
                email:
                  $ref: '#/components/schemas/Email'
                password:
                  type: string
                  format: password
                  minLength: 8
                full_name:
                  type: string
                  minLength: 2
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registered successfully. Please check your email for verification."
                  user_id:
                    $ref: '#/components/schemas/UUID'
        '400':
          $ref: '#/components/responses/ValidationError'
  /auth/logout:
    post:
      summary: User logout
      description: Invalidate the current session
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"
  /auth/refresh:
    post:
      summary: Refresh access token
      description: Get a new access token using refresh token
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: bearer
                  expires_in:
                    type: integer
  /auth/me:
    get:
      summary: Get current user profile
      description: Get the profile of the currently authenticated user
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # Microsoft Entra ID SSO
...